- fix Visual Studio x64 compilation (ambiguous calls to cmath functions)
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@9382 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ plugins/view/PixelOrientedView/PixelOrientedView.cpp
@@ -369,7 +369,7 @@ void PixelOrientedView::initPixelView() {
 
   if (selectedGraphProperties.size()  == 0) return;
 
-  float squareRoot = sqrt(selectedGraphProperties.size());
+  float squareRoot = sqrt(double(selectedGraphProperties.size()));
   const unsigned int N =  (unsigned int) squareRoot + (fmod((float) selectedGraphProperties.size(), squareRoot) == 0. ? 0 : 1);
 
   for (size_t i = 0 ; i < selectedGraphProperties.size() ; ++i) {
